Revised javascript to use jQuery methods wherever possible.
Also rewrote IpAddress's Index view to use IpAddressController (MVC)
instead of IpAddressApiController (WebAPI).  This was not as straight
forward as one might suspect because apparently the
ApplicationController class wraps any return types inside another class
regardless of the class' actual return type.  This makes the original
return type difficult to get at unless the method is rewritten to return
a JsonNetResult of that type.  Once that has occurred the original
return type can be parsed back into a JSON object and its contents acted
upon.  This is unnecessary when using a WebAPI controller because
classes like an Html response object are returned unwrapped, and can be
accessed directly.  Just an FYI